Migrate data object type registry module with bureaucracy module
- change ensure_root() to the ensure_origin_is_active_leader()
- update tests
- add comments

+++ runtime-modules/storage/src/tests/mock.rs
@@ -195,7 +195,7 @@ impl data_object_storage_registry::Trait for Test {
 
 impl members::Trait for Test {
     type Event = MetaEvent;
-    type MemberId = u32;
+    type MemberId = u64;
     type SubscriptionId = u32;
     type PaidTermId = u32;
     type ActorId = u32;
